rename: test 레포 더미 데이터, storage 패키지 구조 변경
