+++ src/test/java/com/sprint/ootd5team/base/storage/S3FileStorageTest.java
@@ -1,4 +1,4 @@
-package com.sprint.ootd5team.storage;
+package com.sprint.ootd5team.base.storage;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -15,7 +15,6 @@ import com.sprint.ootd5team.base.exception.file.FileDeleteFailedException;
 import com.sprint.ootd5team.base.exception.file.FilePermanentSaveFailedException;
 import com.sprint.ootd5team.base.exception.file.FileSaveFailedException;
 import com.sprint.ootd5team.base.exception.file.FileTooLargeException;
-import com.sprint.ootd5team.base.storage.S3FileStorage;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
