+++ src/test/resources/testdata/test-image.png
